server/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
openai-whisper>=20231117
openai>=1.0.0
google-cloud-speech>=2.16.0
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0
aiofiles>=23.0.0